chore: move editor into builder folder
